+++ src/AppRouter.tsx
@@ -7,7 +7,7 @@ import { SplitterProvider } from './components/Splitter/SplitterContext.tsx';
 import { SplitterLayout } from './components/Splitter/SplitterLayout.tsx';
 import { SentryRoutes } from './mount.ts';
 import HeadlampPage from './spaces/mcp/pages/HeadlampPage.tsx';
-import McpPage from './spaces/mcp/pages/McpPage.tsx';
+import ManagedControlPlanePage from './spaces/mcp/pages/ManagedControlPlanePage.tsx';
 import ControlPlanePageV2 from './spaces/controlPlaneV2/pages/ControlPlanePageV2.tsx';
 import ProjectPage from './spaces/onboarding/pages/ProjectPage.tsx';
 import ProjectListView from './views/ProjectList';
@@ -42,7 +42,7 @@ function AppRouter() {
                 />
                 <Route
                   path="projects/:projectName/workspaces/:workspaceName/managedcontrolplane/:controlPlaneName"
-                  element={<McpPage />}
+                  element={<ManagedControlPlanePage />}
                 />
               </Route>
