+++ src/components/ControlPlanes/List/ControlPlaneListWorkspaceGridTile.tsx
@@ -18,7 +18,7 @@ import { DeleteWorkspaceDialog } from '../../Dialogs/KubectlCommandInfo/KubectlD
 import { CopyButton } from '../../Shared/CopyButton.tsx';
 import IllustratedError from '../../Shared/IllustratedError.tsx';
 import { IllustratedBanner } from '../../Ui/IllustratedBanner/IllustratedBanner.tsx';
-import { CreateControlPlaneV2WizardContainer } from '../../Wizards/CreateManagedControlPlane/CreateControlPlaneV2WizardContainer.tsx';
+import { CreateControlPlaneV2WizardContainer } from '../../Wizards/CreateControlPlaneV2/CreateControlPlaneV2WizardContainer.tsx';
 import { CreateManagedControlPlaneWizardContainer } from '../../Wizards/CreateManagedControlPlane/CreateManagedControlPlaneWizardContainer.tsx';
 import { YamlViewButton } from '../../Yaml/YamlViewButton.tsx';
 import { ControlPlaneCard } from '../ControlPlaneCard/ControlPlaneCard.tsx';
