+++ src/spaces/controlPlaneV2/components/Kpi/useEsoQuery.ts
@@ -3,13 +3,13 @@ import { useMemo } from 'react';
 import { z } from 'zod';
 
 import { graphql } from '../../../../types/__generated__/graphql/index.ts';
-import { LandscaperData, LandscaperSchema } from '../../types/Landscaper.ts';
+import { EsoData, EsoSchema } from '../../../mcp/types/Eso.ts';
 
-const GET_LANDSCAPER_QUERY = graphql(`
-  query GetLandscaper($name: String!, $namespace: String) {
-    landscaper_services_open_control_plane_io {
-      v1alpha2 {
-        Landscaper(name: $name, namespace: $namespace) {
+const GET_ESO_QUERY = graphql(`
+  query GetExternalSecretsOperator($name: String!, $namespace: String) {
+    external_secrets_services_open_control_plane_io {
+      v1alpha1 {
+        ExternalSecretsOperator(name: $name, namespace: $namespace) {
           metadata {
             name
             namespace
@@ -18,7 +18,6 @@ const GET_LANDSCAPER_QUERY = graphql(`
             version
           }
           status {
-            phase
             conditions {
               type
               status
@@ -32,20 +31,20 @@ const GET_LANDSCAPER_QUERY = graphql(`
   }
 `);
 
-export function useLandscaperQuery(name?: string, namespace?: string) {
-  const queryResult = useQuery(GET_LANDSCAPER_QUERY, {
+export function useEsoQuery(name?: string, namespace?: string) {
+  const queryResult = useQuery(GET_ESO_QUERY, {
     variables: { name: name ?? '', namespace },
     skip: !name || !namespace,
     notifyOnNetworkStatusChange: true,
   });
 
-  const rawLandscaper = queryResult.data?.landscaper_services_open_control_plane_io?.v1alpha2?.Landscaper;
+  const rawEso = queryResult.data?.external_secrets_services_open_control_plane_io?.v1alpha1?.ExternalSecretsOperator;
 
-  const landscaperData = useMemo<LandscaperData | null>(() => {
-    if (!rawLandscaper) return null;
-    const result = LandscaperSchema.safeParse(rawLandscaper);
+  const esoData = useMemo<EsoData | null>(() => {
+    if (!rawEso) return null;
+    const result = EsoSchema.safeParse(rawEso);
     if (!result.success) {
-      console.warn('[useLandscaperQuery] Validation failed:', z.treeifyError(result.error));
+      console.warn('[useEsoQuery] Validation failed:', z.treeifyError(result.error));
       return null;
     }
     const { spec } = result.data;
@@ -54,10 +53,10 @@ export function useLandscaperQuery(name?: string, namespace?: string) {
       isInstalled: !!version,
       version,
     };
-  }, [rawLandscaper]);
+  }, [rawEso]);
 
   return {
-    landscaperData,
+    esoData,
     isLoading: queryResult.loading,
     error: queryResult.error,
   };
